requirements.txt
torch~=2.7.1
torchvision~=0.22.1
numpy~=2.3.1
PyYAML~=6.0.2
tqdm~=4.67.1
mlflow~=3.1.1
mlflow-skinny~=3.1.1